requirements.txt
Flask
requests
pytest
pydantic
fastapi
uvicorn
python-dotenv
cryptography
PyJWT


